adding changes

+++ interface/renderwindow_bind.cpp
@@ -2,30 +2,30 @@
  * renderwindow_bind.cpp - bindings for Ogre::RenderWindow
  ******************************************************************************
  * This file is part of
- *     __ __              _ 
+ *     __ __              _
  *    / // /_____ ____   (_)
- *   / // // ___// __ \ / / 
- *  / // // /__ / /_/ // /  
- * /_//_/ \___/ \____//_/   
- *                          
+ *   / // // ___// __ \ / /
+ *  / // // /__ / /_/ // /
+ * /_//_/ \___/ \____//_/
+ *
  * Low Level C Ogre Interface (llcoi)
  *
  * See http://code.google.com/p/llcoi/ for more information.
  *
  * Copyright (c) 2011, Llcoi Team
- * 
+ *
  * License: MIT
- * 
+ *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
  * in the Software without restriction, including without limitation the rights
  * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
  * copies of the Software, and to permit persons to whom the Software is
  * furnished to do so, subject to the following conditions:
- * 
+ *
  * The above copyright notice and this permission notice shall be included in
  * all copies or substantial portions of the Software.
- * 
+ *
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
  * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -48,7 +48,7 @@ void render_window_set_visible(RenderWindowHandle window_handle, int visible)
     window->setVisible(visible);
 }
 
-size_t render_window_get_hwnd(RenderWindowHandle window_handle)
+unsigned int render_window_get_hwnd(RenderWindowHandle window_handle)
 {
 	size_t windowHnd = 0;
 	Ogre::RenderWindow* window = reinterpret_cast<Ogre::RenderWindow*>(window_handle);
